// types/CommonInputTextProps.ts
export interface SelectLookupProps {
    label?: string;
    labelText?: string;
    labelClassName?: string;
    name?: any;
    value?: any;
    autoFocus?: boolean;
    className?: string;
    disableClearable?: boolean;
    startAdornments?: any;
    options?: any;
    dropDownWidth?: any;
    required?: boolean;
    textFieldSize?: any;
    error?: any;
    bgColor?: string;
    borderColor?: string;
    focusedBorderColor?: string;
    hoverBorderColor?: string;
    labelColor?: string;
    focusColor?: string;
    helperText?: string;
    disabled?: boolean;
    placeholder?: string;
    onChange?: any;
    formik?: FormikProps<any> | any;
    onInputChange?: any;
}